refactor: hoist model display name mapping into @vm0/core
Both the platform UI and the Slack blocks module had their own copies of
the model-id → friendly-name mapping. Move the single source of truth
into @vm0/core and delete both duplicates. Also extract a couple of
helpers from the Slack org callback route to keep its cyclomatic
complexity within the lint budget.
