+++ turbo/apps/platform/src/views/zero-page/components/settings/provider-dialog-fields.tsx
@@ -12,6 +12,7 @@ import {
   MODEL_PROVIDER_TYPES,
   getAuthMethodsForType,
   getSecretsForAuthMethod,
+  getModelDisplayName,
   getModels,
   getVm0VisibleModels,
   getDefaultModel,
@@ -24,7 +25,6 @@ import {
   getUIDefaultModel,
   getUISecretField,
   getUIAuthMethodLabel,
-  getModelDisplayName,
 } from "./provider-ui-config.ts";
 import { ClaudeCodeSetupPrompt } from "./setup-prompt.tsx";
 import { featureSwitch$ } from "../../../../signals/external/feature-switch.ts";
